refactor: simplify naming slightly
